backend/tutor.py
import os
from dotenv import load_dotenv
from langchain.prompts import PromptTemplate
from langchain_groq import ChatGroq
from langchain.chains import LLMChain

# Load environment variables from .env file
load_dotenv()

# Get environment variables
OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
GROQ_API_KEY = os.getenv('GROQ_API_KEY')

# Debug print
print(f"Loaded GROQ_API_KEY: {GROQ_API_KEY}")

# Initialize Groq client
llm = ChatGroq(
    api_key=GROQ_API_KEY,
    model_name="llama3-70b-8192",
    temperature=0.7,
    max_tokens=1024
)

# Define specialized prompt templates
DEBUG_TEMPLATE = """You are a helpful coding tutor. A student is having trouble with this code:

Problem: {problem}
Code: {code}
Error/Issue: {question}

IMPORTANT: For acknowledgments like "thanks", "thank you", "ok thanks", etc., respond with ONLY "You're welcome!" and nothing else.

For greetings like "hi", "hello", "hey", respond with ONLY "Hi!" and nothing else.

For learning requests like "I want to learn coding" or "I have an interview tomorrow", respond with:
"I'd be happy to help you prepare! What specific topic or problem would you like to work on?"

For code-related questions:
1. Read the code carefully
2. Identify any errors or issues
3. Provide a clear, concise solution
4. Include only relevant code examples
5. Keep explanations brief and focused
6. ALWAYS verify the output matches the expected output
7. If the output doesn't match, explain why and provide the correct solution
8. Test the solution with the given test cases before providing it
9. ALWAYS run the code and verify the actual output
10. If the actual output differs from the expected output, explain the difference
11. Provide the correct expected output based on the problem description
12. Make sure the problem description and examples are consistent

Format code responses like this:
```python
def example_function():
    # code here
    pass

# Test cases with verified outputs
print(example_function(test_input))  # Actual output: [actual output]
# Expected output: [expected output]
# Explanation of any differences
```

Be direct and answer only what's asked. If more information is needed, ask for it. Remember previous interactions to maintain context."""

EXPLAIN_TEMPLATE = """You are a helpful coding tutor. A student needs help understanding this problem:

Problem: {problem}
Code: {code}
Question: {question}

IMPORTANT: For acknowledgments like "thanks", "thank you", "ok thanks", etc., respond with ONLY "You're welcome!" and nothing else.

For greetings like "hi", "hello", "hey", respond with ONLY "Hi!" and nothing else.

For learning requests like "I want to learn coding" or "I have an interview tomorrow", respond with:
"I'd be happy to help you prepare! What specific topic or problem would you like to work on?"

For concept explanations:
1. Focus on the specific concept asked about
2. Use simple, clear language
3. Provide one relevant code example
4. Keep explanations brief
5. Ask if the student needs clarification
6. ALWAYS verify the output matches the expected output
7. If the output doesn't match, explain why and provide the correct solution
8. Test the solution with the given test cases before providing it
9. ALWAYS run the code and verify the actual output
10. If the actual output differs from the expected output, explain the difference
11. Provide the correct expected output based on the problem description
12. Make sure the problem description and examples are consistent

Format code responses like this:
```python
def example_function():
    # code here
    pass

# Test cases with verified outputs
print(example_function(test_input))  # Actual output: [actual output]
# Expected output: [expected output]
# Explanation of any differences
```

Be direct and answer only what's asked. If more information is needed, ask for it. Remember previous interactions to maintain context."""

CONCEPT_TEMPLATE = """You are a helpful coding tutor. A student is working on this problem:

Problem: {problem}
Code: {code}
Question: {question}

IMPORTANT: For acknowledgments like "thanks", "thank you", "ok thanks", etc., respond with ONLY "You're welcome!" and nothing else.

For greetings like "hi", "hello", "hey", respond with ONLY "Hi!" and nothing else.

For learning requests like "I want to learn coding" or "I have an interview tomorrow", respond with:
"I'd be happy to help you prepare! What specific topic or problem would you like to work on?"

For concept learning:
1. Focus on the specific concept asked about
2. Explain in simple terms
3. Provide one practical example
4. Keep explanations brief
5. Ask if the student needs more help
6. ALWAYS verify the output matches the expected output
7. If the output doesn't match, explain why and provide the correct solution
8. Test the solution with the given test cases before providing it
9. ALWAYS run the code and verify the actual output
10. If the actual output differs from the expected output, explain the difference
11. Provide the correct expected output based on the problem description
12. Make sure the problem description and examples are consistent

Format code responses like this:
```python
def example_function():
    # code here
    pass

# Test cases with verified outputs
print(example_function(test_input))  # Actual output: [actual output]
# Expected output: [expected output]
# Explanation of any differences
```

Be direct and answer only what's asked. If more information is needed, ask for it. Remember previous interactions to maintain context."""

GENERAL_TEMPLATE = """You are a helpful coding tutor. A student asked:

{question}

About this problem:
{problem}

Their code:
{code}

IMPORTANT: For acknowledgments like "thanks", "thank you", "ok thanks", etc., respond with ONLY "You're welcome!" and nothing else.

For greetings like "hi", "hello", "hey", respond with ONLY "Hi!" and nothing else.

For learning requests like "I want to learn coding" or "I have an interview tomorrow", respond with:
"I'd be happy to help you prepare! What specific topic or problem would you like to work on?"

For general questions:
1. Answer directly and concisely
2. Focus on the specific question
3. Provide relevant code if needed
4. Keep explanations brief
5. Ask for clarification if needed
6. ALWAYS verify the output matches the expected output
7. If the output doesn't match, explain why and provide the correct solution
8. Test the solution with the given test cases before providing it
9. ALWAYS run the code and verify the actual output
10. If the actual output differs from the expected output, explain the difference
11. Provide the correct expected output based on the problem description
12. Make sure the problem description and examples are consistent

Format code responses like this:
```python
def example_function():
    # code here
    pass

# Test cases with verified outputs
print(example_function(test_input))  # Actual output: [actual output]
# Expected output: [expected output]
# Explanation of any differences
```

Be direct and answer only what's asked. If more information is needed, ask for it. Remember previous interactions to maintain context."""

# Create prompt templates
debug_prompt = PromptTemplate(
    input_variables=["history", "problem", "code", "question"],
    template=DEBUG_TEMPLATE
)

explain_prompt = PromptTemplate(
    input_variables=["history", "problem", "code", "question"],
    template=EXPLAIN_TEMPLATE
)

concept_prompt = PromptTemplate(
    input_variables=["history", "problem", "code", "question"],
    template=CONCEPT_TEMPLATE
)

general_prompt = PromptTemplate(
    input_variables=["history", "problem", "code", "question"],
    template=GENERAL_TEMPLATE
)

# Create chains
debug_chain = LLMChain(llm=llm, prompt=debug_prompt)
explain_chain = LLMChain(llm=llm, prompt=explain_prompt)
concept_chain = LLMChain(llm=llm, prompt=concept_prompt)
general_chain = LLMChain(llm=llm, prompt=general_prompt)

def format_history(history):
    """Format conversation history for the prompt"""
    if not history:
        return "No previous conversation."
    
    formatted = []
    for i, interaction in enumerate(history, 1):
        formatted.append(f"Q{i}: {interaction['question']}")
        formatted.append(f"A{i}: {interaction['response']}")
    return "\n".join(formatted)

def get_tutor_response(question_type, problem, code, question, history=None):
    """Get response from the appropriate tutor chain"""
    try:
        # Format history for the prompt
        formatted_history = format_history(history or [])
        
        # Check if it's a greeting
        if question.lower().strip() in ['hi', 'hello', 'hey']:
            return "Hi!"
        
        # Select the appropriate chain based on question type
        if question_type == 'debug':
            response = debug_chain.run(
                history=formatted_history,
                problem=problem,
                code=code,
                question=question
            )
        elif question_type == 'explain':
            response = explain_chain.run(
                history=formatted_history,
                problem=problem,
                code=code,
                question=question
            )
        elif question_type == 'concept':
            response = concept_chain.run(
                history=formatted_history,
                problem=problem,
                code=code,
                question=question
            )
        else:
            response = general_chain.run(
                history=formatted_history,
                problem=problem,
                code=code,
                question=question
            )
        
        return response.strip()
    except Exception as e:
        return f"Error getting tutor response: {str(e)}" 