backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.4.2
langchain==0.1.12
langchain-groq==0.0.1
gunicorn==21.2.0
pydantic==2.5.2 